[refactor] 폴더 구조 변경, 유저 연결
