+++ src/main/java/com/eat/eat_server/EatServerApplication.java
@@ -2,8 +2,10 @@ package com.eat.eat_server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class EatServerApplication {
 
 	public static void main(String[] args) {
