+++ src/main/java/com/eat/eat_server/domain/logs/domain/Category.java
@@ -1,0 +1,19 @@
+package com.eat.eat_server.domain.logs.domain;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Category {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "category_id")
+    private long id;
+
+    @Column(nullable = false)
+    private String name;
+}
