+++ src/main/java/com/eat/eat_server/domain/user/domain/Role.java
@@ -1,0 +1,5 @@
+package com.eat.eat_server.domain.user.domain;
+
+public enum Role {
+    ROLE_USER, ROLE_ADMIN;
+}
