+++ src/main/java/com/eat/eat_server/domain/user/dto/LoginRequestDto.java
@@ -1,0 +1,15 @@
+package com.eat.eat_server.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class LoginRequestDto {
+    private String email;
+    private String password;
+}
