+++ src/main/java/com/eat/eat_server/domain/user/dto/TokenResponseDto.java
@@ -1,0 +1,16 @@
+package com.eat.eat_server.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TokenResponseDto {
+    private String grantType;
+    private String jwtAccessToken;
+    private String jwtRefreshToken;
+}
