+++ src/main/java/com/eat/eat_server/global/config/SecurityConfig.java
@@ -1,0 +1,75 @@
+package com.eat.eat_server.global.config;
+
+import com.eat.eat_server.domain.user.jwt.JwtAuthenticationFilter;
+import com.eat.eat_server.domain.user.jwt.JwtTokenProvider;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.time.Duration;
+
+@RequiredArgsConstructor
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig {
+
+
+    private final JwtTokenProvider jwtTokenProvider;
+
+    //비밀번호 암호화 처리하는 메서디 제공
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
+    }
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                .cors(cors -> cors.configurationSource(corsConfigurationSource()))
+                .authorizeRequests()
+                .requestMatchers(HttpMethod.OPTIONS, "/**").permitAll() //prefilght request 허용
+                .anyRequest().permitAll();
+
+        http
+                .csrf(csrf -> csrf.disable()) //token을 사용하는 방식이므로 csrf를 disable로 설정
+                .formLogin(formLogin -> formLogin.disable()) //사용자 지정 로그인 로직 구현
+                .httpBasic(httpBasic -> httpBasic.disable());  //http 기본 인증 비활성화
+
+        http.addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class);
+
+        //세션을 사용하지 않으므로 Stateless로 설정
+        http.sessionManagement(sessionManagement -> sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
+
+        //enable Mysql-console
+        http.headers(headers -> headers.frameOptions(options -> options.sameOrigin()));
+
+        return http.build();
+
+    }
+
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration configuration = new CorsConfiguration();
+        configuration.addAllowedOrigin("http://localhost:5173");
+        configuration.addAllowedMethod("*");
+        configuration.addAllowedHeader("*");
+        configuration.setAllowCredentials(true);
+        configuration.setMaxAge(Duration.ofSeconds(3600));
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", configuration);
+        return source;
+    }
+
+}
