--- src/main/java/com/eat/eat_server/user/domain/Role.java
@@ -1,5 +1,0 @@
-package com.eat.eat_server.user.domain;
-
-public enum Role {
-    ROLE_USER, ROLE_ADMIN;
-}
