--- src/main/java/com/eat/eat_server/user/domain/User.java
@@ -1,100 +1,0 @@
-package com.eat.eat_server.user.domain;
-
-import com.eat.eat_server.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-
-
-@Getter
-@Entity
-@NoArgsConstructor
-public class User extends BaseTimeEntity implements UserDetails {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private Long id;
-
-    private String nickname;
-
-    private String email;
-
-    private String password;
-
-    private int age;
-
-    private String gender;
-
-    private Long height;
-
-    private Long weight;
-    private String refreshToken;
-
-    public void encodePassword(PasswordEncoder passwordEncoder){
-        this.password = passwordEncoder.encode(password);
-    }
-
-    public void updateRefreshToken(String refreshToken){
-        this.refreshToken = refreshToken;
-    }
-
-    @Enumerated(EnumType.STRING)
-    private Role role;
-
-
-    @Builder
-    public User (String nickname, String email, String password,
-                 int age, String gender, Long height, Long weight, Role role){
-        this.nickname = nickname;
-        this.email = email;
-        this.password = password;
-        this.age = age;
-        this.gender = gender;
-        this.height = height;
-        this.weight = weight;
-        this.role = role;
-    }
-
-    //userDetail
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        List<GrantedAuthority> auth = new ArrayList<>();
-        auth.add(new SimpleGrantedAuthority(role.name()));
-        return auth;
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
