--- src/main/java/com/eat/eat_server/user/dto/LoginRequestDto.java
@@ -1,15 +1,0 @@
-package com.eat.eat_server.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class LoginRequestDto {
-    private String email;
-    private String password;
-}
