--- src/main/java/com/eat/eat_server/user/dto/TokenResponseDto.java
@@ -1,16 +1,0 @@
-package com.eat.eat_server.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TokenResponseDto {
-    private String grantType;
-    private String jwtAccessToken;
-    private String jwtRefreshToken;
-}
